Move to country module setup
