+++ lib/ibandit.rb
@@ -4,8 +4,9 @@ require 'ibandit/errors'
 require 'ibandit/constants'
 require 'ibandit/iban'
 require 'ibandit/german_details_converter'
-require 'ibandit/swedish_details_converter'
-require 'ibandit/swedish_details_validator'
+require 'ibandit/sweden/local_details_converter'
+require 'ibandit/sweden/validator'
+require 'ibandit/sweden/bank_lookup'
 require 'ibandit/iban_splitter'
 require 'ibandit/iban_assembler'
 require 'ibandit/pseudo_iban_assembler'
